--- services/ui-src/src/components/Logo/index.test.tsx
@@ -1,10 +1,0 @@
-import { Logo } from "./index";
-import { render } from "@testing-library/react";
-
-describe("Test Logo.tsx", () => {
-  test("Check that the QMRLogo renders", () => {
-    const { getByTestId } = render(<Logo />);
-
-    expect(getByTestId("qmr-logo")).toBeVisible();
-  });
-});
