+++ services/ui-src/src/components/index.tsx
@@ -1,6 +1,8 @@
-export * from "./Header";
-export * from "./LocalLogins";
-export * from "./Logo";
-export * from "./wrappers/App";
-export * from "./wrappers/AppRoutes";
-export * from "./wrappers/QueryProvider";
+import { App } from "./wrappers/App";
+import { AppRoutes } from "./wrappers/AppRoutes";
+import { Header } from "./layout/Header";
+import { LoginCognito } from "./logins/LoginCognito";
+import { LoginIDM } from "./logins/LoginIDM";
+import { QueryProvider } from "./wrappers/QueryProvider";
+
+export { App, AppRoutes, Header, LoginCognito, LoginIDM, QueryProvider };
