+++ services/ui-src/src/components/logins/LoginCognito.tsx
@@ -3,34 +3,24 @@ import { Auth } from "aws-amplify";
 // utils
 import { useFormFields } from "../../libs/hooksLib";
 // components
-import {
-  Box,
-  Button,
-  Container,
-  Divider,
-  Heading,
-  Input,
-  Stack,
-} from "@chakra-ui/react";
+import { Button, Divider, Heading, Input, Stack } from "@chakra-ui/react";
 
-interface Props {
-  loginWithIDM: () => void;
-}
-
-const LocalLogin = () => {
+export const LoginCognito = () => {
   const navigate = useNavigate();
   const [fields, handleFieldChange] = useFormFields({
     email: "",
     password: "",
   });
-  async function handleLogin() {
+
+  const handleLogin = async () => {
     try {
       await Auth.signIn(fields.email, fields.password);
       navigate(`/`);
     } catch (error) {
       console.log("Error while logging in.", error); // eslint-disable-line no-console
     }
-  }
+  };
+
   return (
     <Stack>
       <Divider />
@@ -41,23 +31,23 @@ const LocalLogin = () => {
         Email
       </Heading>
       <Input
-        className="field"
-        type="email"
         id="email"
         name="email"
+        type="email"
         value={fields.email}
         onChange={handleFieldChange}
+        className="field"
       />
       <Heading mb="2" size="sm">
         Password
       </Heading>
       <Input
-        className="field"
-        type="password"
         id="password"
         name="password"
+        type="password"
         value={fields.password}
         onChange={handleFieldChange}
+        className="field"
       />
       <Button
         colorScheme="teal"
@@ -70,24 +60,5 @@ const LocalLogin = () => {
         Login with Cognito
       </Button>
     </Stack>
-  );
-};
-
-export const LocalLogins = ({ loginWithIDM }: Props) => {
-  return (
-    <Container maxW="sm" h="full" my="auto">
-      <Box textAlign="center" mb="6">
-        <Heading mb="2" size="md" alignSelf="center">
-          Developer Login{" "}
-        </Heading>
-        <Divider />
-      </Box>
-      <Stack spacing={8}>
-        <Button colorScheme="teal" onClick={loginWithIDM} isFullWidth>
-          Login with IDM
-        </Button>
-        <LocalLogin />
-      </Stack>
-    </Container>
   );
 };
