+++ services/ui-src/src/components/logins/LoginIDM.tsx
@@ -1,0 +1,18 @@
+// components
+import { Box, Button, Divider, Heading } from "@chakra-ui/react";
+
+interface Props {
+  loginWithIDM: () => void;
+}
+
+export const LoginIDM = ({ loginWithIDM }: Props) => (
+  <Box textAlign="center" mb="6">
+    <Heading mb="2" size="md" alignSelf="center">
+      Login with IDM
+    </Heading>
+    <Divider />
+    <Button colorScheme="teal" onClick={loginWithIDM} isFullWidth>
+      Login with IDM
+    </Button>
+  </Box>
+);
