+++ services/ui-src/src/components/wrappers/AppRoutes.tsx
@@ -1,7 +1,7 @@
 import { Route, Routes } from "react-router-dom";
 import * as Views from "views";
 
-const AppRoutes = () => (
+export const AppRoutes = () => (
   <main id="main-wrapper">
     <Routes>
       <Route path="/" element={<Views.Home />} />
@@ -9,5 +9,3 @@ const AppRoutes = () => (
     </Routes>
   </main>
 );
-
-export default AppRoutes;
