+++ services/ui-src/src/components/wrappers/QueryProvider.tsx
@@ -1,11 +1,7 @@
 import { QueryClient, QueryClientProvider } from "react-query";
 
-const queryClient = new QueryClient();
-
-const QueryProvider = ({ children }: any) => {
-  return (
-    <QueryClientProvider client={queryClient}>{children}</QueryClientProvider>
-  );
-};
-
-export default QueryProvider;
+export const QueryProvider = ({ children }: any) => (
+  <QueryClientProvider client={new QueryClient()}>
+    {children}
+  </QueryClientProvider>
+);
