+++ services/ui-src/src/views/NotFound/index.tsx
@@ -1,9 +1,9 @@
-import * as CUI from "@chakra-ui/react";
+import { Flex, Text } from "@chakra-ui/react";
 
 export function NotFound() {
   return (
-    <CUI.Flex h="100%" justifyContent="center" py="12">
-      <CUI.Text data-testid="not-found">Oops, page not found :( </CUI.Text>
-    </CUI.Flex>
+    <Flex h="100%" justifyContent="center" py="12">
+      <Text data-testid="not-found">Oops, page not found :( </Text>
+    </Flex>
   );
 }
